+++ Runtime/SDK/IOS/Domain/IOSNonRenewingPurchase.cs
@@ -1,6 +1,7 @@
 #if UNITY_IOS
 
 using Apphud.Unity.Domain;
+using Newtonsoft.Json;
 using UnityEngine.Scripting;
 
 namespace Apphud.Unity.IOS.Domain
@@ -16,6 +17,8 @@ namespace Apphud.Unity.IOS.Domain
 
     internal sealed class IOSApphudNonRenewingPurchase : ApphudNonRenewingPurchase
     {
+        internal IOSApphudNonRenewingPurchase(string json) : this(JsonConvert.DeserializeObject<IOSApphudNonRenewingPurchaseJson>(json)) { }
+
         internal IOSApphudNonRenewingPurchase(IOSApphudNonRenewingPurchaseJson json)
         {
             ProductId = json.productId;